fix bugs

+++ Controllers/LessonController.cs
@@ -32,6 +32,14 @@ namespace ELProject.Controllers
             var InstructorId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             if (InstructorId == null) return Unauthorized("User Not Authenticated");
 
+            var section = await unitOfWork.Sections.GetByIdAsync(dto.SectionId);
+            if (section == null)
+                return NotFound("Section not found");
+
+            var course = await unitOfWork.Courses.GetByIdAsync(section.CourseId);
+            if (course.UserId != InstructorId)
+                return Forbid();
+
             var url = await fileService.UploadFileAsync(dto.File, dto.Type);
 
             if (string.IsNullOrEmpty(url))
@@ -103,6 +111,11 @@ namespace ELProject.Controllers
             if (lesson == null)
                 return NotFound("Lesson Not Found");
 
+            var section = await unitOfWork.Sections.GetByIdAsync(lesson.SectionId);
+            var course = await unitOfWork.Courses.GetByIdAsync(section.CourseId);
+            if (course.UserId != InstructorId)
+                return Forbid();
+
             if (dto.File != null)
             {
                 if (!string.IsNullOrEmpty(lesson.FileUrl))
@@ -136,6 +149,11 @@ namespace ELProject.Controllers
 
             if (lesson == null)
                 return NotFound("Lesson Not Found");
+
+            var section = await unitOfWork.Sections.GetByIdAsync(lesson.SectionId);
+            var course = await unitOfWork.Courses.GetByIdAsync(section.CourseId);
+            if (course.UserId != InstructorId)
+                return Forbid();
 
             if (!string.IsNullOrEmpty(lesson.FileUrl))
                 await fileService.DeleteFileAsync(lesson.FileUrl, lesson.Type);
